backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
motor==3.6.0
pymongo==4.9.2

# Vector store
chromadb==0.5.23

# Document processing
pypdf==5.1.0
python-docx==1.1.2
tiktoken==0.8.0

# Embeddings & LLM
openai==1.58.1
httpx==0.28.1

# Data validation & settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.1
uuid6==2024.7.10
